Pending add sale and tickets to the database at the same time

+++ database/migrations/2017_06_24_184550_create_settings_table.php
@@ -26,7 +26,7 @@ class CreateSettingsTable extends Migration
             $table->decimal('members_weekend')->nullable();
             $table->decimal('members_matinee')->nullable();
             $table->decimal('members_special_event')->nullable();
-            $table->decimal('tax')->nullable()->default(8.25);
+            $table->float('tax')->nullable()->default(8.25);
             $table->timestamps();
         });
     }
